+++ packages/common/src/util/clientSupportsFallBack.ts
@@ -1,5 +1,5 @@
 import type { Command } from "../types/command/command.types";
 
-export function useFallback(command: Command): boolean {
+export function clientSupportsFallback(command: Command): boolean {
   return command.version >= 7;
 }
