+++ packages/cursorless-engine/src/api/CursorlessEngineApi.ts
@@ -39,7 +39,7 @@ export interface CommandApi {
    * Runs a command.  This is the core of the Cursorless engine.
    * @param command The command to run
    */
-  runCommand(command: Command): Promise<CommandResponse>;
+  runCommand(command: Command): Promise<CommandResponse | unknown>;
 
   /**
    * Designed to run commands that come directly from the user.  Ensures that
