+++ packages/cursorless-engine/src/cursorlessEngine.ts
@@ -1,12 +1,10 @@
 import {
   Command,
-  CommandResponse,
   CommandServerApi,
   FileSystem,
   Hats,
   IDE,
   ScopeProvider,
-  useFallback,
 } from "@cursorless/common";
 import {
   CommandRunnerDecorator,
@@ -89,8 +87,7 @@ export function createCursorlessEngine(
       },
 
       async runCommandSafe(...args: unknown[]) {
-        const command = ensureCommandShape(args);
-        const response = await runCommand(
+        return runCommand(
           treeSitter,
           commandServerApi,
           debug,
@@ -100,9 +97,8 @@ export function createCursorlessEngine(
           languageDefinitions,
           rangeUpdater,
           commandRunnerDecorators,
-          command,
+          ensureCommandShape(args),
         );
-        return unwrapCommandResponse(command, response);
       },
     },
     scopeProvider: createScopeProvider(
@@ -121,19 +117,6 @@ export function createCursorlessEngine(
       commandRunnerDecorators.push(decorator);
     },
   };
-}
-
-async function unwrapCommandResponse(
-  command: Command,
-  response: CommandResponse,
-): Promise<CommandResponse | unknown> {
-  if (useFallback(command)) {
-    return response;
-  }
-  if ("returnValue" in response) {
-    return response.returnValue;
-  }
-  return undefined;
 }
 
 function createScopeProvider(
